fix some folder reorganization
