+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/models/dataservice/modelparts/semantics/Transition.java
@@ -1,14 +1,12 @@
-package software.uncharted.terarium.hmiserver.models.dataservice.modelparts;
+package software.uncharted.terarium.hmiserver.models.dataservice.modelparts.semantics;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
 
-import com.fasterxml.jackson.annotation.JsonAlias;
-import com.fasterxml.jackson.annotation.JsonSetter;
 import software.uncharted.terarium.hmiserver.annotations.TSOptional;
 
 import java.util.List;
-import software.uncharted.terarium.hmiserver.models.dataservice.modelparts.metadata.VariableStatement;
+import software.uncharted.terarium.hmiserver.models.dataservice.modelparts.ModelGrounding;
 
 @Data
 @Accessors(chain = true)
