+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/models/dataservice/modelparts/semantics/TypingSemantics.java
@@ -1,4 +1,4 @@
-package software.uncharted.terarium.hmiserver.models.dataservice.modelparts;
+package software.uncharted.terarium.hmiserver.models.dataservice.modelparts.semantics;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
@@ -6,7 +6,6 @@ import lombok.experimental.Accessors;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.List;
-import software.uncharted.terarium.hmiserver.models.dataservice.modelparts.metadata.VariableStatement;
 
 @Data
 @Accessors(chain = true)
